Renamed remaining endpoints.py to model.py for consistency
